+++ pandora/validation/validation.py
@@ -138,7 +138,7 @@ class AbstractValidation:
         :param img_left: left Datset image containing :
 
             - im: 2D (row, col) or 3D (band_im, row, col) xarray.DataArray float32
-            - disparity (optional): 3D (disp, row, col) xarray.DataArray int16
+            - disparity (optional): 3D (disp, row, col) xarray.DataArray float32
             - msk (optional): 2D (row, col) xarray.DataArray int16
             - classif (optional): 3D (band_classif, row, col) xarray.DataArray int16
             - segm (optional): 2D (row, col) xarray.DataArray int16
@@ -146,7 +146,7 @@ class AbstractValidation:
         :param img_right: right Dataset image containing :
 
             - im: 2D (row, col) or 3D (band_im, row, col) xarray.DataArray float32
-            - disparity (optional): 3D (disp, row, col) xarray.DataArray int16
+            - disparity (optional): 3D (disp, row, col) xarray.DataArray float32
             - msk (optional): 2D (row, col) xarray.DataArray int16
             - classif (optional): 3D (band_classif, row, col) xarray.DataArray int16
             - segm (optional): 2D (row, col) xarray.DataArray int16
@@ -247,7 +247,7 @@ class CrossCheckingAccurate(AbstractValidation):
         :param img_left: left Datset image containing :
 
                 - im: 2D (row, col) or 3D (band_im, row, col) xarray.DataArray float32
-                - disparity (optional): 3D (disp, row, col) xarray.DataArray int16
+                - disparity (optional): 3D (disp, row, col) xarray.DataArray float32
                 - msk (optional): 2D (row, col) xarray.DataArray int16
                 - classif (optional): 3D (band_classif, row, col) xarray.DataArray int16
                 - segm (optional): 2D (row, col) xarray.DataArray int16
@@ -255,7 +255,7 @@ class CrossCheckingAccurate(AbstractValidation):
         :param img_right: right Dataset image containing :
 
                 - im: 2D (row, col) or 3D (band_im, row, col) xarray.DataArray float32
-                - disparity (optional): 3D (disp, row, col) xarray.DataArray int16
+                - disparity (optional): 3D (disp, row, col) xarray.DataArray float32
                 - msk (optional): 2D (row, col) xarray.DataArray int16
                 - classif (optional): 3D (band_classif, row, col) xarray.DataArray int16
                 - segm (optional): 2D (row, col) xarray.DataArray int16
